+++ example/test_driver/integration_test.dart
@@ -1,0 +1,12 @@
+import 'dart:io';
+
+import 'package:integration_test/integration_test_driver_extended.dart';
+
+Future<void> main() => integrationDriver(
+  onScreenshot: (name, bytes, [args]) async {
+    final file = File('build/device-debug/$name.png');
+    await file.parent.create(recursive: true);
+    await file.writeAsBytes(bytes);
+    return true;
+  },
+);